Add a placeholder test, flake8 should ignore a `.venv`
Initial CI pytest calls fail with a non-zero exit code if there are
absolutely no tests. A placeholder test module and function are added to
be at least something for pytest to see and run so that there is a
success return.

Also found flake8 in CI checking into the local virtual environment
which is of course not desired, so add the configured directory to the
exclude list.